pyproject.toml
[project]
name = "tangerine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.3.0",
    "pyqt6>=6.9.1",
    "pytorch-lightning>=2.5.5",
    "scikit-image>=0.25.2",
    "scipy>=1.16.2",
    "segmentation-models-pytorch>=0.5.0",
    "shapely>=2.1.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
